add funcs

+++ app/Http/Controllers/CustomerController.php
@@ -9,10 +9,17 @@ use Inertia\Inertia;
 
 class CustomerController extends Controller
 {
+    public function getCustomers(Request $request) {
+        $customers = Customer::query();
+        $customers = $customers->latest('id');
+        if ($request->search) {
+            $customers = $customers->where('info->cmnd->number', 'like', '%'.$request->search.'%');
+        }
+        return response()->json($customers->paginate(10));
+    }
+
     public function index() {
-        return Inertia::render('Customer/Index', [
-            'customers' => Customer::latest('id')->paginate(10),
-        ]);
+        return Inertia::render('Customer/Index');
     }
 
     public function create() {
@@ -36,8 +43,14 @@ class CustomerController extends Controller
                         'time' => '',
                     ],
                     'assets' => '',
-                    'notes' => '',
-                    'files' => [],
+                    'notes' => [
+                        'saler' => [],
+                        'appraiser' => [],
+                    ],
+                    'files' => [
+                        'saler' => [],
+                        'appraiser' => [],
+                    ],
                     'relative_contacts' => [[
                         'phone' => '',
                         'relative' => ''
@@ -54,13 +67,23 @@ class CustomerController extends Controller
         $validated = $request->validated();
         unset($validated['info']['files']);
         $customer = Customer::create($validated);
-        $customer->addMultipleMediaFromRequest(['info.files'])->each(function ($fileAdder) {
-            $fileAdder->toMediaCollection('customers');
-        });
+
+        if (count($request->info['files']['saler'])) {
+            $customer->addMultipleMediaFromRequest(['info.files.saler'])->each(function ($fileAdder) {
+                $fileAdder->toMediaCollection('saler_files');
+            });
+        }
+        if (count($request->info['files']['appraiser'])) {
+            $customer->addMultipleMediaFromRequest(['info.files.appraiser'])->each(function ($fileAdder) {
+                $fileAdder->toMediaCollection('appraiser_files');
+            });
+        }
+
         return Inertia::render('Customer/Index');
     }
 
     public function edit(Customer $customer) {
+        $customer->setRelation('payments', $customer->payments()->paginate(10));
         return Inertia::render('Customer/Modify', [
             'customer' => $customer,
         ]);
@@ -77,10 +100,11 @@ class CustomerController extends Controller
 
     public function update(CustomerRequest $request, Customer $customer) {
         $customer->update($request->validated());
-        return Inertia::render('Customer/Index');
+        return redirect()->route('customers.index');
     }
 
     public function destroy(Customer $customer) {
+        $this->authorize('delete', $customer);
         $customer->delete();
         return response()->noContent();
     }
